# perf: replace `StringBuilder` with `ValueListBuilder<char>`