--- a/Src/CSharpier.Core/CSharp/SyntaxPrinter/CSharpierIgnore.cs
+++ b/Src/CSharpier.Core/CSharp/SyntaxPrinter/CSharpierIgnore.cs
@@ -1,6 +1,8 @@
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.RegularExpressions;
 using CSharpier.Core.DocTypes;
+using CSharpier.Core.Utilities;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -63,6 +65,7 @@ internal static partial class CSharpierIgnore
             && HasIgnoreComment(syntaxNode);
     }
 
+    [SkipLocalsInit]
     public static List<Doc> PrintNodesRespectingRangeIgnore<T>(
         SyntaxList<T> list,
         PrintingContext context
@@ -70,14 +73,14 @@ internal static partial class CSharpierIgnore
         where T : SyntaxNode
     {
         var statements = new List<Doc>();
-        var unFormattedCode = new StringBuilder();
+        var unFormattedCode = new ValueListBuilder<char>(stackalloc char[64]);
         var printUnformatted = false;
 
         foreach (var node in list)
         {
             if (Token.HasLeadingCommentMatching(node, IgnoreEndRegex))
             {
-                statements.Add(unFormattedCode.ToString().Trim());
+                statements.Add(unFormattedCode.AsSpan().Trim().ToString());
                 unFormattedCode.Clear();
                 printUnformatted = false;
             }
@@ -98,8 +101,10 @@ internal static partial class CSharpierIgnore
 
         if (unFormattedCode.Length > 0)
         {
-            statements.Add(unFormattedCode.ToString().Trim());
+            statements.Add(unFormattedCode.AsSpan().Trim().ToString());
         }
+
+        unFormattedCode.Dispose();
 
         return statements;
     }