--- a/Src/CSharpier.Core/CSharp/SyntaxPrinter/MembersWithForcedLines.cs
+++ b/Src/CSharpier.Core/CSharp/SyntaxPrinter/MembersWithForcedLines.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using System.Runtime.CompilerServices;
 using System.Text;
 using CSharpier.Core.DocTypes;
 using CSharpier.Core.Utilities;
@@ -10,6 +11,7 @@ namespace CSharpier.Core.CSharp.SyntaxPrinter;
 
 internal static class MembersWithForcedLines
 {
+    [SkipLocalsInit]
     public static List<Doc> Print<T>(
         CSharpSyntaxNode node,
         IReadOnlyList<T> members,
@@ -24,7 +26,7 @@ internal static class MembersWithForcedLines
             result.Add(Doc.HardLine);
         }
 
-        var unFormattedCode = new StringBuilder();
+        var unFormattedCode = new ValueListBuilder<char>(stackalloc char[64]);
         var printUnformatted = false;
         var lastMemberForcedBlankLine = false;
         for (var memberIndex = 0; memberIndex < members.Count; memberIndex++)
@@ -35,7 +37,7 @@ internal static class MembersWithForcedLines
             if (Token.HasLeadingCommentMatching(member, CSharpierIgnore.IgnoreEndRegex))
             {
                 skipAddingLineBecauseIgnoreEnded = true;
-                result.Add(unFormattedCode.ToString().Trim());
+                result.Add(unFormattedCode.AsSpan().Trim().ToString());
                 unFormattedCode.Clear();
                 printUnformatted = false;
             }
@@ -229,8 +231,10 @@ internal static class MembersWithForcedLines
 
         if (unFormattedCode.Length > 0)
         {
-            result.Add(unFormattedCode.ToString().Trim());
+            result.Add(unFormattedCode.AsSpan().ToString().Trim());
         }
+
+        unFormattedCode.Dispose();
 
         return result;
     }