--- a/Src/CSharpier.Core/CSharp/SyntaxPrinter/SeparatedSyntaxList.cs
+++ b/Src/CSharpier.Core/CSharp/SyntaxPrinter/SeparatedSyntaxList.cs
@@ -1,3 +1,4 @@
+using System.Runtime.CompilerServices;
 using System.Text;
 using CSharpier.Core.DocTypes;
 using CSharpier.Core.Utilities;
@@ -34,6 +35,7 @@ internal static class SeparatedSyntaxList
     // the names above aren't totally accurate
     // sometimes there are trailing commas with calls to Print (some patterns do that)
     // and if you pass null to PrintWithTrailingComma it won't add a trailing comma if there isn't one
+    [SkipLocalsInit]
     private static Doc Print<T>(
         in SeparatedSyntaxList<T> list,
         Func<T, PrintingContext, Doc> printFunc,
@@ -48,7 +50,7 @@ internal static class SeparatedSyntaxList
             list.Count <= 3
                 ? new ValueListBuilder<Doc>([null, null, null, null, null, null, null, null])
                 : new ValueListBuilder<Doc>(list.Count * 3);
-        var unFormattedCode = new StringBuilder();
+        var unFormattedCode = new ValueListBuilder<char>(stackalloc char[64]);
         var printUnformatted = false;
         for (var x = startingIndex; x < list.Count; x++)
         {
@@ -56,7 +58,7 @@ internal static class SeparatedSyntaxList
 
             if (Token.HasLeadingCommentMatching(member, CSharpierIgnore.IgnoreEndRegex))
             {
-                docs.Append(unFormattedCode.ToString().Trim());
+                docs.Append(unFormattedCode.AsSpan().Trim().ToString());
                 unFormattedCode.Clear();
                 printUnformatted = false;
             }
@@ -74,7 +76,8 @@ internal static class SeparatedSyntaxList
                 unFormattedCode.Append(CSharpierIgnore.PrintWithoutFormatting(member, context));
                 if (x < list.SeparatorCount)
                 {
-                    unFormattedCode.AppendLine(list.GetSeparator(x).Text);
+                    unFormattedCode.Append(list.GetSeparator(x).Text);
+                    unFormattedCode.Append(Environment.NewLine);
                 }
 
                 continue;
@@ -144,10 +147,11 @@ internal static class SeparatedSyntaxList
 
         if (unFormattedCode.Length > 0)
         {
-            docs.Append(unFormattedCode.ToString().Trim());
+            docs.Append(unFormattedCode.AsSpan().Trim().ToString());
         }
 
         var output = Doc.Concat(ref docs);
+        unFormattedCode.Dispose();
         docs.Dispose();
 
         return output;