@TimothyMakkison 2025-07-18
In retrospect I think I was being overly paranoid. This is a net positive for memory usage with a negligible improvement to speed. The timing of the benchmarks naturally varies a little